muliple base convert to single base and identifiers table seprated
